[WIP] - Cleanup and refactoring
