+++ geonode/security/utils.py
@@ -16,6 +16,8 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
 #########################################################################
+import copy
+import json
 import logging
 from itertools import chain
 
@@ -30,7 +32,18 @@ from guardian.shortcuts import get_objects_for_user
 
 from geonode.groups.models import GroupProfile
 from geonode.groups.conf import settings as groups_settings
-from geonode.security.permissions import DATASET_EDIT_DATA_PERMISSIONS
+from geonode.security.permissions import (
+    PermSpecCompact,
+    VIEW_PERMISSIONS,
+    ADMIN_PERMISSIONS,
+    SERVICE_PERMISSIONS,
+    DOWNLOAD_PERMISSIONS,
+    DOWNLOADABLE_RESOURCES,
+    BASIC_MANAGE_PERMISSIONS,
+    DATASET_ADMIN_PERMISSIONS,
+    DATASET_EDIT_DATA_PERMISSIONS,
+    DATASET_EDIT_STYLE_PERMISSIONS,
+    DATA_EDITABLE_RESOURCES_SUBTYPES)
 
 logger = logging.getLogger(__name__)
 
@@ -256,3 +269,329 @@ def get_user_visible_groups(user, include_public_invite: bool = False):
         [metadata_author_groups.append(item) for item in all_metadata_author_groups
             if item not in metadata_author_groups]
     return metadata_author_groups
+
+
+class AdvancedSecurityWorkflowManager:
+
+    @staticmethod
+    def is_group_private_mode():
+        return settings.GROUP_PRIVATE_RESOURCES
+
+    @staticmethod
+    def is_manager_publish_mode():
+        return settings.RESOURCE_PUBLISHING
+
+    @staticmethod
+    def is_admin_moderate_mode():
+        return settings.ADMIN_MODERATE_UPLOADS
+
+    @staticmethod
+    def is_auto_publishing_workflow():
+        """
+          **AUTO PUBLISHING**
+            - `RESOURCE_PUBLISHING = False`
+            - `ADMIN_MODERATE_UPLOADS = False`
+
+            - When user creates a resource:
+              - OWNER gets all the owner permissions (publish resource included)
+              - ANONYMOUS can view and download
+            - No change to the Group Manager is applied
+        """
+        return not settings.RESOURCE_PUBLISHING and not settings.ADMIN_MODERATE_UPLOADS
+
+    @staticmethod
+    def is_simple_publishing_workflow():
+        """
+          **SIMPLE PUBLISHING**
+            - `RESOURCE_PUBLISHING = True` (Autopublishing is disabled)
+            - `ADMIN_MODERATE_UPLOADS = False`
+
+            - When user creates a resource:
+              - OWNER gets all the owner permissions (`publish_resource` and `change_resourcebase_permissions` INCLUDED)
+              - Group MANAGERS of the user's groups will get the owner permissions (`publish_resource` EXCLUDED)
+              - Group MEMBERS of the user's groups will get the `view_resourcebase`, `download_resourcebase` permission
+              - ANONYMOUS can not view and download if the resource is not published
+
+            - When resource has a group assigned:
+              - OWNER gets all the owner permissions (`publish_resource` and `change_resourcebase_permissions` INCLUDED)
+              - Group MANAGERS of the *resource's group* will get the owner permissions (`publish_resource` EXCLUDED)
+              - Group MEMBERS of the *resource's group* will get the `view_resourcebase`, `download_resourcebase` permission
+        """
+        return settings.RESOURCE_PUBLISHING and not settings.ADMIN_MODERATE_UPLOADS
+
+    @staticmethod
+    def is_advanced_workflow():
+        """
+          **ADVANCED WORKFLOW**
+            - `RESOURCE_PUBLISHING = True`
+            - `ADMIN_MODERATE_UPLOADS = True`
+
+            - When user creates a resource:
+              - OWNER gets all the owner permissions (`publish_resource` and `change_resourcebase_permissions` EXCLUDED)
+              - Group MANAGERS of the user's groups will get the owner permissions (`publish_resource` INCLUDED)
+              - Group MEMBERS of the user's groups will get the `view_resourcebase`, `download_resourcebase` permission
+              - ANONYMOUS can not view and download if the resource is not published
+
+            - When resource has a group assigned:
+              - OWNER gets all the owner permissions (`publish_resource` and `change_resourcebase_permissions` EXCLUDED)
+              - Group MANAGERS of the resource's group will get the owner permissions (`publish_resource` INCLUDED)
+              - Group MEMBERS of the resource's group will get the `view_resourcebase`, `download_resourcebase` permission
+        """
+        return settings.RESOURCE_PUBLISHING and settings.ADMIN_MODERATE_UPLOADS
+
+    @staticmethod
+    def is_simplified_workflow():
+        """
+          **SIMPLIFIED WORKFLOW**
+            - `RESOURCE_PUBLISHING = False`
+            - `ADMIN_MODERATE_UPLOADS = True`
+
+            - **NOTE**: Is it even possibile? when the resource is automatically published, can it be un-published?
+            If this combination is not allowed, we should either stop the process when reading the settings or log a warning and force a safe combination.
+
+            - When user creates a resource:
+              - OWNER gets all the owner permissions (`publish_resource` and `change_resourcebase_permissions` INCLUDED)
+              - Group MANAGERS of the user's groups will get the owner permissions (`publish_resource` INCLUDED)
+              - Group MEMBERS of the user's group will get the `view_resourcebase`, `download_resourcebase` permission
+              - ANONYMOUS can view and download
+        """
+        return not settings.RESOURCE_PUBLISHING and settings.ADMIN_MODERATE_UPLOADS
+
+    @staticmethod
+    def assignable_perm_condition(perm, resource_type):
+        _assignable_perm_policy_condition = (perm in DOWNLOAD_PERMISSIONS and resource_type in DOWNLOADABLE_RESOURCES) or \
+            (perm in DATASET_EDIT_DATA_PERMISSIONS and resource_type in DATA_EDITABLE_RESOURCES_SUBTYPES) or \
+            (perm not in (DOWNLOAD_PERMISSIONS + DATASET_EDIT_DATA_PERMISSIONS))
+        logger.debug(f" perm: {perm} - resource_type: {resource_type} --> assignable: {_assignable_perm_policy_condition}")
+        return _assignable_perm_policy_condition
+
+    @staticmethod
+    def get_instance(uuid: str):
+        from geonode.base.models import ResourceBase
+
+        _resources = ResourceBase.objects.filter(uuid=uuid)
+        _exists = _resources.count() == 1
+        if _exists:
+            return _resources.get()
+        return None
+
+    @staticmethod
+    def set_group_manager_permissions_according_to_workflow(user, group, role):
+
+        def _handle_perms(perms=None):
+            '''
+            Internally the set_permissions function will automatically handle the permissions
+            that needs to be assigned to re resource.
+            Background at: https://github.com/GeoNode/geonode/pull/8145
+            If the user is demoted, we assign by default at least the view and the download permission
+            to the resource
+            '''
+            queryset = (
+                get_objects_for_user(
+                    user,
+                    ["base.view_resourcebase", "base.change_resourcebase"],
+                    any_perm=True)
+                .filter(group=group.group)
+                .exclude(owner=user)
+            )
+            # A.F.: By including 'group.resources()' here, we will look also for resources
+            #       having permissions related to the current 'group' and not only the ones assigned
+            #       to the 'group' through the metadata settings.
+            # _resources = set([_r for _r in queryset.iterator()] + [_r for _r in group.resources()])
+            _resources = queryset.iterator()
+            for _r in _resources:
+                perm_spec = _r.get_all_level_info()
+                if perms:
+                    if "users" not in perm_spec:
+                        perm_spec["users"] = {}
+                    perm_spec["users"][user] = perms
+                # Let's the ResourceManager finally decide which are the correct security settings to apply
+                _r.set_permissions(perm_spec)
+
+        if not AdvancedSecurityWorkflowManager.is_auto_publishing_workflow():
+            if role and role == "manager":
+                _handle_perms(perms=VIEW_PERMISSIONS + DOWNLOAD_PERMISSIONS + BASIC_MANAGE_PERMISSIONS)
+            else:
+                _handle_perms(perms=VIEW_PERMISSIONS + DOWNLOAD_PERMISSIONS)
+
+    @staticmethod
+    def get_owner_permissions(uuid: str, /, instance=None, members: list = [], perm_spec: dict = {"users": {}, "groups": {}}) -> dict:
+        """compute a perm_spec to assing all admin permissions to the owner"""
+        _resource = instance or AdvancedSecurityWorkflowManager.get_instance(uuid)
+        _perm_spec = copy.deepcopy(perm_spec)
+
+        if _resource:
+            _resource = _resource.get_real_instance()
+            if _resource.polymorphic_ctype:
+                # Owner & Manager Admin Perms
+                admin_perms = VIEW_PERMISSIONS + ADMIN_PERMISSIONS
+                if _resource.polymorphic_ctype.name in DOWNLOADABLE_RESOURCES:
+                    admin_perms += DOWNLOAD_PERMISSIONS
+                for perm in admin_perms:
+                    if not settings.RESOURCE_PUBLISHING and not settings.ADMIN_MODERATE_UPLOADS:
+                        _prev_perm = _perm_spec["users"].get(_resource.owner, []) if "users" in _perm_spec else []
+                        _perm_spec["users"][_resource.owner] = set.union(perms_as_set(_prev_perm), perms_as_set(perm))
+                    elif perm not in {'change_resourcebase_permissions', 'publish_resourcebase'}:
+                        _prev_perm = _perm_spec["users"].get(_resource.owner, []) if "users" in _perm_spec else []
+                        _perm_spec["users"][_resource.owner] = set.union(perms_as_set(_prev_perm), perms_as_set(perm))
+                    if members:
+                        for user in members:
+                            _prev_perm = _perm_spec["users"].get(user, []) if "users" in _perm_spec else []
+                            _perm_spec["users"][user] = set.union(perms_as_set(_prev_perm), perms_as_set(perm))
+
+                # Set the GeoFence Owner Rule
+                if _resource.polymorphic_ctype.name == 'dataset':
+                    DATA_EDIT_PERMISSIONS = []
+                    if _resource.get_real_instance().subtype == 'vector':
+                        DATA_EDIT_PERMISSIONS = DATASET_ADMIN_PERMISSIONS
+                    elif _resource.get_real_instance().subtype == 'raster':
+                        DATA_EDIT_PERMISSIONS = DATASET_EDIT_STYLE_PERMISSIONS
+                    for perm in DATA_EDIT_PERMISSIONS:
+                        _prev_perm = _perm_spec["users"].get(_resource.owner, []) if "users" in _perm_spec else []
+                        _perm_spec["users"][_resource.owner] = set.union(perms_as_set(_prev_perm), perms_as_set(perm))
+                        if members:
+                            for user in members:
+                                _prev_perm = _perm_spec["users"].get(user, []) if "users" in _perm_spec else []
+                                _perm_spec["users"][user] = set.union(perms_as_set(_prev_perm), perms_as_set(perm))
+
+                if _resource.polymorphic_ctype.name == 'service':
+                    for perm in SERVICE_PERMISSIONS:
+                        _prev_perm = _perm_spec["users"].get(_resource.owner, []) if "users" in _perm_spec else []
+                        _perm_spec["users"][_resource.owner] = set.union(perms_as_set(_prev_perm), perms_as_set(perm))
+                        if members:
+                            for user in members:
+                                _prev_perm = _perm_spec["users"].get(user, []) if "users" in _perm_spec else []
+                                _perm_spec["users"][user] = set.union(perms_as_set(_prev_perm), perms_as_set(perm))
+        return _perm_spec
+
+    @staticmethod
+    def get_workflow_permissions(uuid: str, /, instance=None, permissions: dict = {}) -> dict:
+        """
+        Adapts the provided "perm_spec" accordingly to the following schema:
+
+                          |  N/PUBLISHED   | PUBLISHED
+          --------------------------------------------
+            N/APPROVED    |     GM/OWR     |     -
+            APPROVED      |   registerd    |    all
+          --------------------------------------------
+
+        It also adds Group Managers as "editors" to the "perm_spec" in the case:
+         - The Advanced Workflow has been enabled
+         - The Group Managers are missing from the provided "perm_spec"
+
+        General Rules:
+         - OWNER can always publish, except in the ADVANCED WORKFLOW
+         - When the user is OWNER and MANAGER, the MANAGER wins!
+         - Group MANAGERS have publish privs only when `ADMIN_MODERATE_UPLOADS` is True (no DATA EDIT perms assigned by default)
+         - Group MEMBERS have always access to the resource, except for the AUTOPUBLISH, where everybody has access to it.
+        """
+        _resource = instance or AdvancedSecurityWorkflowManager.get_instance(uuid)
+
+        _permissions = None
+        if permissions:
+            if PermSpecCompact.validate(permissions):
+                _permissions = PermSpecCompact(copy.deepcopy(permissions), _resource).extended
+            else:
+                _permissions = copy.deepcopy(permissions)
+
+        if _resource:
+            perm_spec = _permissions or copy.deepcopy(_resource.get_all_level_info())
+
+            # Sanity checks
+            if isinstance(perm_spec, str):
+                perm_spec = json.loads(perm_spec)
+
+            if "users" not in perm_spec:
+                perm_spec["users"] = {}
+            elif isinstance(perm_spec["users"], list):
+                _users = {}
+                for _item in perm_spec["users"]:
+                    _users[_item[0]] = _item[1]
+                perm_spec["users"] = _users
+
+            if "groups" not in perm_spec:
+                perm_spec["groups"] = {}
+            elif isinstance(perm_spec["groups"], list):
+                _groups = {}
+                for _item in perm_spec["groups"]:
+                    _groups[_item[0]] = _item[1]
+                perm_spec["groups"] = _groups
+
+            # Make sure we're dealing with "Profile"s and "Group"s...
+            perm_spec = _resource.fixup_perms(perm_spec)
+            _resource_type = _resource.resource_type or _resource.polymorphic_ctype.name
+
+            if not AdvancedSecurityWorkflowManager.is_auto_publishing_workflow():
+                # compute advanced workflow permissions
+                if _resource_type not in DOWNLOADABLE_RESOURCES:
+                    view_perms = VIEW_PERMISSIONS
+                else:
+                    view_perms = VIEW_PERMISSIONS + DOWNLOAD_PERMISSIONS
+
+                admin_perms = ADMIN_PERMISSIONS.copy()
+
+                admin_nopub_perms = ADMIN_PERMISSIONS.copy()
+                admin_nopub_perms.remove('publish_resourcebase')
+
+                admin_restricted_perms = ADMIN_PERMISSIONS.copy()
+                admin_restricted_perms.remove('publish_resourcebase')
+                admin_restricted_perms.remove('change_resourcebase_permissions')
+
+                anonymous_group = Group.objects.get(name='anonymous')
+                registered_members_group_name = groups_settings.REGISTERED_MEMBERS_GROUP_NAME
+                user_groups = Group.objects.filter(
+                    name__in=_resource.owner.groupmember_set.all().values_list("group__slug", flat=True))
+                member_group_perm, group_managers = _resource.get_group_managers(user_groups)
+
+                # Handle the owner perms
+                if _resource.owner:
+                    prev_perms = perm_spec['users'].get(_resource.owner, []) if isinstance(perm_spec['users'], dict) else []
+                    if AdvancedSecurityWorkflowManager.is_advanced_workflow():
+                        perm_spec['users'][_resource.owner] = list(
+                                    set(prev_perms + view_perms + admin_restricted_perms))
+                    else:
+                        perm_spec['users'][_resource.owner] = list(
+                                    set(prev_perms + view_perms + admin_perms))
+
+                # Handle the Group Managers perms
+                if group_managers:
+                    for group_manager in group_managers:
+                        prev_perms = perm_spec['users'].get(group_manager, []) if isinstance(perm_spec['users'], dict) else []
+                        # AF: Should be a manager being able to change the dataset data and style too by default?
+                        #     For the time being let's give to the manager "management" perms only.
+                        # if _resource.polymorphic_ctype.name == 'layer':
+                        #     perm_spec['users'][group_manager] = list(
+                        #         set(prev_perms + view_perms + ADMIN_PERMISSIONS + LAYER_ADMIN_PERMISSIONS))
+                        # else:
+                        if AdvancedSecurityWorkflowManager.is_simple_publishing_workflow():
+                            perm_spec['users'][group_manager] = list(
+                                set(prev_perms + view_perms + admin_nopub_perms))
+                        else:
+                            perm_spec['users'][group_manager] = list(
+                                set(prev_perms + view_perms + admin_perms))
+
+                # Handle the Group Members perms
+                if member_group_perm:
+                    for gr, perm in member_group_perm['groups'].items():
+                        if gr != anonymous_group and gr.name != registered_members_group_name:
+                            prev_perms = perm_spec['groups'].get(gr, []) if isinstance(perm_spec['groups'], dict) else []
+                            perm_spec['groups'][gr] = list(set(prev_perms + perm))
+
+                # Handle the Anoymous and other members perms
+                if _resource.is_approved:
+                    if getattr(groups_settings, 'AUTO_ASSIGN_REGISTERED_MEMBERS_TO_REGISTERED_MEMBERS_GROUP_NAME', False):
+                        registered_members_group = Group.objects.get(name=registered_members_group_name)
+                        prev_perms = perm_spec['groups'].get(registered_members_group, []) if isinstance(perm_spec['groups'], dict) else []
+                        perm_spec['groups'][registered_members_group] = list(set(prev_perms + view_perms))
+                    else:
+                        prev_perms = perm_spec['groups'].get(anonymous_group, []) if isinstance(perm_spec['groups'], dict) else []
+                        perm_spec['groups'][anonymous_group] = list(set(prev_perms + view_perms))
+
+                if _resource.is_published:
+                    prev_perms = perm_spec['groups'].get(anonymous_group, []) if isinstance(perm_spec['groups'], dict) else []
+                    perm_spec['groups'][anonymous_group] = list(set(prev_perms + view_perms))
+            else:
+                # default permissions for resource owner
+                perm_spec = AdvancedSecurityWorkflowManager.get_owner_permissions(_resource.uuid, perm_spec=perm_spec)
+
+        return perm_spec
